fixed the bugs in the range checking code after running the tests again

+++ mixl/tests/testthat/test_maxlikelihood.R
@@ -3,7 +3,7 @@ context("Test maxlikelihood function")
 data("Train", package="mlogit")
 head(Train, 3)
 Train$ID <- Train$id
-Train$CHOICE <- as.numeric(Train$choice)
+Train$CHOICE <- as.integer(Train$choice)
 
 Nindividuals <- length(unique(Train$id))
